Freemarker native support #223
