+++ docs/modules/ROOT/partials/reference/components/freemarker.adoc
@@ -2,11 +2,11 @@
 // This file was generated by camel-quarkus-maven-plugin:update-extension-doc-page
 :cq-artifact-id: camel-quarkus-freemarker
 :cq-artifact-id-base: freemarker
-:cq-native-supported: false
-:cq-status: Preview
+:cq-native-supported: true
+:cq-status: Stable
 :cq-deprecated: false
 :cq-jvm-since: 1.1.0
-:cq-native-since: n/a
+:cq-native-since: 1.8.0
 :cq-camel-part-name: freemarker
 :cq-camel-part-title: Freemarker
 :cq-camel-part-description: Transform messages using FreeMarker templates.
