+++ extensions/freemarker/runtime/src/main/doc/configuration.adoc
@@ -1,44 +1,3 @@
-// Do not edit directly!
-// This file was generated by camel-quarkus-maven-plugin:update-extension-doc-page
-= Freemarker
-:cq-artifact-id: camel-quarkus-freemarker
-:cq-native-supported: true
-:cq-status: Stable
-:cq-description: Transform messages using FreeMarker templates.
-:cq-deprecated: false
-:cq-jvm-since: 1.1.0
-:cq-native-since: 1.8.0
-
-[.badges]
-[.badge-key]##JVM since##[.badge-supported]##1.1.0## [.badge-key]##Native since##[.badge-supported]##1.8.0##
-
-Transform messages using FreeMarker templates.
-
-== What's inside
-
-* xref:{cq-camel-components}::freemarker-component.adoc[Freemarker component], URI syntax: `freemarker:resourceUri`
-
-Please refer to the above link for usage and configuration details.
-
-== Maven coordinates
-
-[source,xml]
-----
-<dependency>
-    <groupId>org.apache.camel.quarkus</groupId>
-    <artifactId>camel-quarkus-freemarker</artifactId>
-</dependency>
-----
-
-Check the xref:user-guide/index.adoc[User guide] for more information about writing Camel Quarkus applications.
-
-== allowContextMapAll option in native mode
-
-The `allowContextMapAll` option is not supported in native mode as it requires reflective access to security sensitive camel core classes such as
-`CamelContext` & `Exchange`. This is considered a security risk and thus access to the feature is not provided by default.
-
-== Additional Camel Quarkus configuration
-
 == Class path resources
 
 For the FreeMarker templates loaded from class path to work flawlessly in native mode, you need to list all your
@@ -123,4 +82,3 @@ public class MyRoutes extends RouteBuilder {
     }
 }
 ----
-
