+++ extensions/pom.xml
@@ -115,6 +115,7 @@
         <module>file-watch</module>
         <module>flatpack</module>
         <module>fop</module>
+        <module>freemarker</module>
         <module>ftp</module>
         <module>geocoder</module>
         <module>git</module>
