+++ integration-tests/freemarker/src/main/java/org/apache/camel/quarkus/component/freemarker/it/MyPerson.java
