+++ integration-tests/freemarker/src/main/resources/AppleTemplate.ftl
