+++ integration-tests/freemarker/src/main/resources/folder/subfolder/templates/BodyAsDomainObject.ftl
