+++ integration-tests/freemarker/src/main/resources/org/apache/camel/component/freemarker/example.ftl
